Renamed demonata to src to prevent name collisions Fixed EmptyRoom class Updated lint rules Removed utils.py (curses leftover)
