--- demonata/demonata.py
@@ -1,7 +1,0 @@
-"""Main game file"""
-from controller import Controller
-
-if __name__ == "__main__":
-    print "Welcome to Demonata\n"
-    GC = Controller()
-    GC.start()
