+++ src/view/initial_view.py
